backend/requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
joblib
scikit-learn
numpy
python-dotenv